requirement.txt
absl-py==2.3.1
astunparse==1.6.3
audioread==3.0.1
brainflow==5.16.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
flatbuffers==25.2.10
fonttools==4.58.5
freetype-py==2.5.1
gast==0.6.0
glcontext==3.0.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.1
keras-nightly==3.10.0.dev2025050203
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.43.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
moderngl==5.12.0
msgpack==1.1.1
multipledispatch==1.0.0
namex==0.1.0
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.21.6
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
pyrr==0.10.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
tabulate==0.9.0
tb-nightly==2.19.0a20250218
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_nightly==2.20.0.dev20250425
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
